requirements.txt
py-cord[voice]
python-dotenv
google-generativeai
openai